+++ src/mocks/VatMock.sol
@@ -31,18 +31,18 @@ contract VatMock {
     }
 
     struct Ilk {
-        uint256 Art;   // Total Normalised Debt     [wad]
-        uint256 rate;  // Accumulated Rates         [ray]
-        uint256 spot;  // Price with Safety Margin  [ray]
-        uint256 line;  // Debt Ceiling              [rad]
-        uint256 dust;  // Urn Debt Floor            [rad]
+        uint256 Art; // Total Normalised Debt     [wad]
+        uint256 rate; // Accumulated Rates         [ray]
+        uint256 spot; // Price with Safety Margin  [ray]
+        uint256 line; // Debt Ceiling              [rad]
+        uint256 dust; // Urn Debt Floor            [rad]
     }
 
-    mapping(bytes32 => Ilk)                         public ilks;
+    mapping(bytes32 => Ilk) public ilks;
     mapping(address => mapping(address => uint256)) public can;
-    mapping(bytes32 => mapping(address => Urn))     public urns;
+    mapping(bytes32 => mapping(address => Urn)) public urns;
     mapping(bytes32 => mapping(address => uint256)) public gem;
-    mapping(address => uint256)                     public dai;
+    mapping(address => uint256) public dai;
 
     constructor() {
         wards[msg.sender] = 1;
@@ -51,10 +51,10 @@ contract VatMock {
     function init(bytes32 ilk) external auth {
         require(ilks[ilk].rate == 0, "Vat/ilk-already-init");
         ilks[ilk].rate = 10 ** 27;
-        ilks[ilk].line = 20_000_000 * 10 ** 45;        
+        ilks[ilk].line = 20_000_000 * 10 ** 45;
     }
 
-    function file(bytes32 ilk, bytes32 what, uint data) external auth {
+    function file(bytes32 ilk, bytes32 what, uint256 data) external auth {
         if (what == "line") ilks[ilk].line = data;
         else revert("Vat/file-unrecognized-param");
     }
@@ -81,8 +81,8 @@ contract VatMock {
 
         urn.ink = _add(urn.ink, dink);
         urn.art = _add(urn.art, dart);
-        ilk.Art =_add(ilk.Art, dart);
-        int dtab = int(ilk.rate) * dart;
+        ilk.Art = _add(ilk.Art, dart);
+        int256 dtab = int256(ilk.rate) * dart;
 
         require(dart <= 0 || ilk.Art * ilk.rate <= ilk.line, "Vat/ceiling-exceeded");
 
@@ -91,7 +91,7 @@ contract VatMock {
         dai[w] = dtab >= 0 ? dai[w] + uint256(dtab) : dai[w] - uint256(-dtab);
 
         urns[i][u] = urn;
-        ilks[i]    = ilk;
+        ilks[i] = ilk;
     }
 
     function grab(bytes32 i, address u, address v, address, int256 dink, int256 dart) external auth {
@@ -99,27 +99,29 @@ contract VatMock {
         Ilk storage ilk = ilks[i];
 
         urn.ink = _add(urn.ink, dink);
-        urn.art =_add(urn.art, dart);
+        urn.art = _add(urn.art, dart);
         ilk.Art = _add(ilk.Art, dart);
         gem[i][v] = _sub(gem[i][v], dink);
     }
 
-    function fold(bytes32 i, int rate) external auth {
+    function fold(bytes32 i, int256 rate) external auth {
         Ilk storage ilk = ilks[i];
         ilk.rate = _add(ilk.rate, rate);
     }
 
-    /** Math */
-    function _add(uint x, int y) internal pure returns (uint z) {
+    /**
+     * Math
+     */
+    function _add(uint256 x, int256 y) internal pure returns (uint256 z) {
         z = y >= 0 ? x + uint256(y) : x - uint256(-y);
     }
 
-    function _sub(uint256 x, int y) internal pure returns (uint z) {
-        z = y >= 0 ? x - uint256(y) :x + uint256(-y);
+    function _sub(uint256 x, int256 y) internal pure returns (uint256 z) {
+        z = y >= 0 ? x - uint256(y) : x + uint256(-y);
     }
 
-    function _mul(uint x, int y) internal pure returns (int z) {
-        z = int(x) * y;
-        require(y == 0 || z / y == int(x), "VatMock/mul-overflow");
+    function _mul(uint256 x, int256 y) internal pure returns (int256 z) {
+        z = int256(x) * y;
+        require(y == 0 || z / y == int256(x), "VatMock/mul-overflow");
     }
 }
